Handling long filenames of mesos-slave's sandbox dirs
rsyslog crashes if the filename that it has to monitor exceeds 200 chars. To work around that, we create a symlink to the task dir of mesos sandbox and add that path as the file to be watched in rsyslog. This should safely keep it under the 200 char file path limit